requirements.txt
streamlit
pandas
python-dotenv
langchain
langchain-groq
groq
langchain-experimental
tabulate
